🎉 Rename code-snippets to nocloud 🎉 Time to graduate the code to the **real** NoCloud name!
