+++ nocloud/cloud_init/CloudInitUserData.py
