+++ nocloud/cloud_init/models/__init__.py
