+++ nocloud/cloud_init/models/base_generator.py
