+++ nocloud/cloud_init/models/meta_data.py
