+++ nocloud/cloud_init/models/user_data.py
