+++ nocloud/cloud_init/models/vendor_data.py
