+++ nocloud/domain_instance/detach_disk.py
