+++ nocloud/domain_instance/get_domain_uuid.py
