+++ nocloud/domain_instance/get_vm_status.py
