+++ nocloud/domain_instance/start_vm.py
@@ -1,5 +1,5 @@
 import libvirt
-from code_snippets.exception import FuckYouLibvirtError
+from nocloud.exception import FuckYouLibvirtError
 # Parameters
 LIBVIRT_URI = 'qemu+ssh://chong601@10.102.0.5/system'
 DOMAIN_NAME = 'vm-ubuntu-focal-lxd-1'
@@ -8,8 +8,7 @@ client = libvirt.open(LIBVIRT_URI)
 
 try:
     domain = client.lookupByName(DOMAIN_NAME)
-    domain.shutdown()
-
+    domain.create()
 except libvirt.libvirtError:
     a = FuckYouLibvirtError()
     print(a.getLibvirtRawErrorData())
