+++ nocloud/storage-pool/autostart_storage_pool.py
