+++ nocloud/storage-pool/create_storage_pool.py
