+++ nocloud/storage-pool/deactivate_storage_pool.py
