+++ nocloud/storage-pool/delete_storage_pool.py
