+++ nocloud/storage-pool/get_storage_pool_by_name.py
