+++ nocloud/storage-pool/get_storage_pool_path.py
