+++ nocloud/storage-pool/get_storage_pool_xml.py
