+++ nocloud/storage-pool/list_storage_pool_by_name.py
