+++ nocloud/storage-volume/delete_vol_from_pool.py
