+++ nocloud/storage-volume/get_storage_volume_by_name.py
