+++ nocloud/storage-volume/get_storage_volume_xml.py
