+++ nocloud/templates/cloudinit/user-data
