+++ nocloud/templates/cloudinit/vendor-data
