+++ nocloud/templates/xmldef/centosdomainxmldef.xml
