+++ nocloud/templates/xmldef/diskxmldef.xml
