+++ nocloud/templates/xmldef/poolxmldef.xml
